requirements.txt
streamlit
python-dotenv
sentinelhub
numpy
shapely
geojson
rasterio
scikit-image
matplotlib
